Swapped views -> init for secrets, services,root
