+++ ftplugin/k8s_deployments.lua
@@ -3,7 +3,7 @@ local api = vim.api
 local deplyoment_view = require("kubectl.views.deployments")
 local hl = require("kubectl.actions.highlight")
 local pod_view = require("kubectl.views.pods")
-local root_view = require("kubectl.views.root.views")
+local root_view = require("kubectl.views.root")
 local string_util = require("kubectl.utils.string")
 local view = require("kubectl.views")
