+++ ftplugin/k8s_nodes.lua
@@ -1,6 +1,6 @@
 local node_view = require("kubectl.views.nodes")
 local string_util = require("kubectl.utils.string")
-local root_view = require("kubectl.views.root.views")
+local root_view = require("kubectl.views.root")
 local hl = require("kubectl.actions.highlight")
 local api = vim.api
