+++ ftplugin/k8s_services.lua
@@ -1,6 +1,6 @@
-local secret_view = require("kubectl.views.secrets.views")
+local secret_view = require("kubectl.views.secrets")
 local string_util = require("kubectl.utils.string")
-local root_view = require("kubectl.views.root.views")
+local root_view = require("kubectl.views.root")
 local hl = require("kubectl.actions.highlight")
 local api = vim.api
