+++ lua/kubectl/views/services/init.lua
@@ -1,53 +1,26 @@
+local ResourceBuilder = require("kubectl.resourcebuilder")
+local definition = require("kubectl.views.services.definition")
+
 local M = {}
-local time = require("kubectl.utils.time")
 
-local function getPorts(ports)
-  local string_ports = ""
-  for index, value in ipairs(ports) do
-    string_ports = string_ports .. value.port .. "/" .. value.protocol
-
-    if index < #ports then
-      string_ports = string_ports .. ","
-    end
-  end
-  return string_ports
+function M.Services()
+  ResourceBuilder:new("services", { "get", "services", "-A", "-o=json" })
+    :fetch()
+    :decodeJson()
+    :process(definition.processRow)
+    :prettyPrint(definition.getHeaders)
+    :addHints({
+      { key = "<d>", desc = "describe" },
+    }, true, true)
+    :setFilter(FILTER)
+    :display("k8s_services", "Services")
 end
 
---TODO: Get externalip
-local function getExternalIP(spec)
-  return ""
-end
-
-function M.processRow(rows)
-  local data = {}
-  for _, row in pairs(rows.items) do
-    local pod = {
-      namespace = row.metadata.namespace,
-      name = row.metadata.name,
-      type = row.type,
-      clusterip = row.spec.clusterIP,
-      externalip = getExternalIP(row.spec),
-      ports = getPorts(row.spec.ports),
-      age = time.since(row.metadata.creationTimestamp),
-    }
-
-    table.insert(data, pod)
-  end
-  return data
-end
-
-function M.getHeaders()
-  local headers = {
-    "NAMESPACE",
-    "NAME",
-    "TYPE",
-    "CLUSTERIP",
-    "EXTERNALIP",
-    "PORTS",
-    "AGE",
-  }
-
-  return headers
+function M.ServiceDesc(namespace, name)
+  ResourceBuilder:new("desc", { "describe", "svc", name, "-n", namespace })
+    :fetch()
+    :splitData()
+    :displayFloat("k8s_svc_desc", name, "yaml")
 end
 
 return M
